Category and Movie Handlers Unit Tests

+++ src/MovieLibrary.Core/Features/Handlers/RemoveCategoryHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using MovieLibrary.Core.Exceptions;
 using MovieLibrary.Core.Features.Commands;
 using MovieLibrary.Data;
 using System.Threading;
@@ -18,8 +19,11 @@ namespace MovieLibrary.Core.Features.Handlers
         public async Task<Unit> Handle(RemoveCategoryCommand request, CancellationToken cancellationToken)
         {
             var category = await _unitOfWork.CategoryRepository.GetByIdAsync(request.CategoryId);
-            await _unitOfWork.CategoryRepository.RemoveAsync(category);
 
+            if (category == null)
+                throw new CategoryException($"Category wtih id {request.CategoryId} does not exist.");
+
+            await _unitOfWork.CategoryRepository.RemoveAsync(category);
             await _unitOfWork.SaveAsync(cancellationToken);
 
             return Unit.Value;
